Added "double buffering" to the BaseBlockRecordIndex
There seems to be a problem particularly with explosions where
if a block is restore during the same tick it's being blown up
it does not actually get restored.

This works around this issue by restoring the block over the course
of two ticks. While it's technically still possible for the issue to
occur, this should make it significantly more rare.