Actually fix the snow block despawning problem in frostborn
This appears to be due to restoring too early during the boot process
triggering some kind of adverse effect where snow blocks are despawned
and replaced by snow balls -- as if they were broken by a shovel.

My best guess is that this is an upstream bug (Spigot),
or it's caused by a plugin, which I haven't been able to determine.
I checked craftbook's snow mechanic, we both did not have it enabled,
and there seems to be no code in craftbook that would
trigger this exact effect.