Improve the grave yard experience at nighttime
- Players are no longer blinded outside of the temple
- Zombies no longer spawn directly on players outside of the temple, instead they spawn at nearby graves
- The phantom potion now blocks zombies spawning at nearby graves

Additionally at an API level a new event has been added for when a player "disturbs" a grave, triggering a zombie to spawn there.
The grave yard also now keeps track of headstones by position instead of location for a memory and CPU savings (as these no longer need cloned for API reasons). This was especially important given the potentially large number of headstones locations that could need cloned for the new event.